Fixed some details. Add map point empty files.

+++ orb_slam/3rd_part/dbow2/FORB.cpp
@@ -97,7 +97,7 @@ double FORB::distance(const FORB::TDescriptor &a,
       (sizeof(uint64_t) - 1) * CHAR_BIT;
   }
   
-  return ret;
+  return (double)ret;
   
   // // If uint64_t is not defined in your system, you can try this 
   // // portable approach
@@ -171,14 +171,14 @@ void FORB::toMat32F(const std::vector<TDescriptor> &descriptors,
     
     for(int j = 0; j < C; ++j, p += 8)
     {
-      p[0] = (desc[j] & (1 << 7) ? 1 : 0);
-      p[1] = (desc[j] & (1 << 6) ? 1 : 0);
-      p[2] = (desc[j] & (1 << 5) ? 1 : 0);
-      p[3] = (desc[j] & (1 << 4) ? 1 : 0);
-      p[4] = (desc[j] & (1 << 3) ? 1 : 0);
-      p[5] = (desc[j] & (1 << 2) ? 1 : 0);
-      p[6] = (desc[j] & (1 << 1) ? 1 : 0);
-      p[7] = desc[j] & (1);
+      p[0] = (float)(desc[j] & (1 << 7) ? 1 : 0);
+      p[1] = (float)(desc[j] & (1 << 6) ? 1 : 0);
+      p[2] = (float)(desc[j] & (1 << 5) ? 1 : 0);
+      p[3] = (float)(desc[j] & (1 << 4) ? 1 : 0);
+      p[4] = (float)(desc[j] & (1 << 3) ? 1 : 0);
+      p[5] = (float)(desc[j] & (1 << 2) ? 1 : 0);
+      p[6] = (float)(desc[j] & (1 << 1) ? 1 : 0);
+      p[7] = (float)(desc[j] & (1));
     }
   } 
 }
@@ -209,14 +209,14 @@ void FORB::toMat32F(const cv::Mat &descriptors, cv::Mat &mat)
   {
     for(int j = 0; j < C; ++j, p += 8)
     {
-      p[0] = (desc[j] & (1 << 7) ? 1 : 0);
-      p[1] = (desc[j] & (1 << 6) ? 1 : 0);
-      p[2] = (desc[j] & (1 << 5) ? 1 : 0);
-      p[3] = (desc[j] & (1 << 4) ? 1 : 0);
-      p[4] = (desc[j] & (1 << 3) ? 1 : 0);
-      p[5] = (desc[j] & (1 << 2) ? 1 : 0);
-      p[6] = (desc[j] & (1 << 1) ? 1 : 0);
-      p[7] = desc[j] & (1);
+      p[0] = (float)(desc[j] & (1 << 7) ? 1 : 0);
+      p[1] = (float)(desc[j] & (1 << 6) ? 1 : 0);
+      p[2] = (float)(desc[j] & (1 << 5) ? 1 : 0);
+      p[3] = (float)(desc[j] & (1 << 4) ? 1 : 0);
+      p[4] = (float)(desc[j] & (1 << 3) ? 1 : 0);
+      p[5] = (float)(desc[j] & (1 << 2) ? 1 : 0);
+      p[6] = (float)(desc[j] & (1 << 1) ? 1 : 0);
+      p[7] = (float)(desc[j] & (1));
     }
   } 
 }
